+++ hikka/dispatcher.py
@@ -416,14 +416,29 @@ class CommandDispatcher:
                 or whitelist_modules
                 and f"{str(utils.get_chat_id(message))}.{func.__self__.__module__}"
                 not in whitelist_modules
-                or getattr(func, "no_commands", False)
-                and await self._handle_command(event, watcher=True)
-                or getattr(func, "out", False)
-                and not getattr(message, "out", True)
-                or getattr(func, "in", False)
-                and getattr(message, "out", True)
-                or getattr(func, "only_messages", False)
-                and not isinstance(message, types.Message)
+                or (
+                    getattr(func, "no_commands", False)
+                    and await self._handle_command(event, watcher=True)
+                )
+                or (
+                    getattr(func, "only_commands", False)
+                    and not await self._handle_command(event, watcher=True)
+                )
+                or (getattr(func, "out", False) and not getattr(message, "out", True))
+                or (getattr(func, "in", False) and getattr(message, "out", True))
+                or (
+                    getattr(func, "only_messages", False)
+                    and not isinstance(message, types.Message)
+                )
+                or (
+                    getattr(func, "editable", False)
+                    and (
+                        getattr(message, "fwd_from", False)
+                        or not getattr(message, "out", False)
+                        or getattr(message, "sticker", False)
+                        or getattr(message, "via_bot_id", False)
+                    )
+                )
             ):
                 logging.debug(f"Ignored watcher of module {modname}")
                 continue